requirements.txt
asttokens==2.0.4
atomicwrites==1.4.0
attrs==20.3.0
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cymem==2.0.5
decorator==4.4.2
en-core-web-sm==3.0.0
executing==0.5.4
filelock==3.0.12
flake8==3.9.0
future==0.18.2
graphviz==0.16
icecream==2.1.0
idna==2.10
importlib-metadata==3.7.3
iniconfig==1.1.1
install==1.3.4
jellyfish==0.8.2
Jinja2==2.11.3
joblib==1.0.1
keybert==0.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
murmurhash==1.0.5
networkx==2.5
nltk==3.5
numpy==1.20.1
packaging==20.9
pandas==1.2.3
pathy==0.4.0
Pillow==8.1.2
pip==21.0.1
pke==1.8.1
pluggy==0.13.1
preshed==3.0.5
py==1.10.0
pycodestyle==2.7.0
pydantic==1.7.3
pyflakes==2.3.0
Pygments==2.8.1
pyparsing==2.4.7
pystempel==1.2.0
pytest==6.2.2
pytextrank==3.1.0
python-dateutil==2.8.1
python-rake==1.5.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
segtok==1.5.10
sentencepiece==0.1.95
sentence-transformers==0.4.1.2
setuptools==54.1.2
six==1.15.0
sklearn==0.0
smart-open==3.0.0
sortedcontainers==2.3.0
spacy==3.0.5
spacy-legacy==3.0.1
srsly==2.4.0
tabulate==0.8.9
thinc==8.0.2
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.2
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.59.0
transformers==4.4.1
typer==0.3.2
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.4
wasabi==0.8.2
yake==0.4.6
zipp==3.4.1
